requirements.txt
flask>=3.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
google-genai>=1.30.0
numpy>=1.26.0
